im-util: Create separate module for public traits
