+++ eyeball-im-util/src/vector.rs
@@ -3,6 +3,7 @@
 mod filter;
 mod limit;
 mod ops;
+mod traits;
 
 use eyeball_im::VectorDiff;
 use futures_core::Stream;
@@ -13,23 +14,8 @@ use self::ops::{
 pub use self::{
     filter::{Filter, FilterMap},
     limit::{EmptyLimitStream, Limit},
+    traits::VectorDiffContainer,
 };
-
-/// Abstraction over stream items that the adapters in this module can deal
-/// with.
-pub trait VectorDiffContainer: VectorDiffContainerOps<Self::Element> {
-    /// The element type of the [`Vector`][imbl::Vector] that diffs are being
-    /// handled for.
-    type Element;
-}
-
-impl<T> VectorDiffContainer for VectorDiff<T> {
-    type Element = T;
-}
-
-impl<T> VectorDiffContainer for Vec<VectorDiff<T>> {
-    type Element = T;
-}
 
 /// Type alias for extracting the element type from a stream of
 /// [`VectorDiffContainer`]s.
